requirements.txt
aws-cdk-lib==2.80.0
constructs==10.1.0
boto3>=1.28.0
pydantic>=2.0.0